requirements.txt
appnope==0.1.0
beautifulsoup4==4.6.0
bs4==0.0.1
cachetools==2.0.0
certifi==2017.4.17
chardet==3.0.4
coverage==4.4.1
decorator==4.1.2
defusedxml==0.5.0
dill==0.2.7.1
Django==1.11.3
django-extensions==1.8.1
factory-boy==2.8.1
Faker==0.7.18
future==0.16.0
gapic-google-cloud-datastore-v1==0.15.3
gapic-google-cloud-error-reporting-v1beta1==0.15.3
gapic-google-cloud-logging-v2==0.91.3
gapic-google-cloud-pubsub-v1==0.15.4
gapic-google-cloud-spanner-admin-database-v1==0.15.3
gapic-google-cloud-spanner-admin-instance-v1==0.15.3
gapic-google-cloud-spanner-v1==0.15.3
gapic-google-cloud-speech-v1==0.15.3
google-api-python-client==1.6.2
google-auth==1.0.1
google-auth-httplib2==0.0.2
google-auth-oauthlib==0.1.0
google-cloud==0.26.1
google-cloud-bigquery==0.25.0
google-cloud-bigtable==0.25.0
google-cloud-core==0.25.0
google-cloud-datastore==1.1.0
google-cloud-dns==0.25.0
google-cloud-error-reporting==0.25.1
google-cloud-language==0.25.0
google-cloud-logging==1.1.0
google-cloud-monitoring==0.25.0
google-cloud-pubsub==0.26.0
google-cloud-resource-manager==0.25.0
google-cloud-runtimeconfig==0.25.0
google-cloud-spanner==0.25.0
google-cloud-speech==0.26.0
google-cloud-storage==1.2.0
google-cloud-translate==0.25.0
google-cloud-videointelligence==0.25.0
google-cloud-vision==0.25.0
google-gax==0.15.13
google-resumable-media==0.2.1
googleapis-common-protos==1.5.2
grpc-google-iam-v1==0.11.1
grpcio==1.4.0
httplib2==0.10.3
idna==2.5
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
lxml==3.8.0
oauth2client==4.1.2
oauthlib==2.0.2
pexpect==4.2.1
pickleshare==0.7.4
pkgconfig==1.2.2
pluggy==0.4.0
ply==3.8
postgres==2.2.1
prompt-toolkit==1.0.14
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-error-reporting-v1beta1==0.15.3
proto-google-cloud-logging-v2==0.91.3
proto-google-cloud-pubsub-v1==0.15.4
proto-google-cloud-spanner-admin-database-v1==0.15.3
proto-google-cloud-spanner-admin-instance-v1==0.15.3
proto-google-cloud-spanner-v1==0.15.3
proto-google-cloud-speech-v1==0.15.3
protobuf==3.3.0
psycopg2==2.7.3
ptyprocess==0.5.2
py==1.4.34
pyasn1==0.3.1
pyasn1-modules==0.0.10
pycryptodomex==3.4.6
Pygments==2.2.0
pyjwkest==1.3.6
PyJWT==1.5.2
pytest==3.1.3
pytest-cov==2.5.1
python-crontab==2.2.3
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.2
requests==2.18.2
requests-oauthlib==0.8.0
rsa==3.4.2
simplegeneric==0.8.1
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
tox==2.7.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
virtualenv==15.1.0
wcwidth==0.1.7
